opacity support for gifTextureView added

+++ src/main/jni/bitmap.c
@@ -65,7 +65,7 @@ Java_pl_droidsonroids_gif_GifInfoHandle_renderFrame(JNIEnv *env, jclass __unused
         return 0;
     }
     DDGifSlurp(info, true);
-    const uint_fast16_t frameDuration = getBitmap((argb *) pixels, info);
+    const uint_fast32_t frameDuration = getBitmap((argb *) pixels, info);
     unlockPixels(env, jbitmap);
     return calculateInvalidationDelay(info, renderStartTime, frameDuration);
 }
